Update ci.yml
Update ci.yml due to coverage data being computed correctly on an
Ubuntu EC2 instance but not the actions runner as it included both
jobs.c files. It should only compute coverage on the jobs.c file in the
build dir since that has static keywords stripped out for testing.